HOT FIX: mistakenly swapped the arguments to calculate p-value, along with misconception of p-value, I was not aware of this bug.
P-value of "mean < X" is NOT a chance of getting mean < X. It IS the chance of getting null hypothesis (the one that contradicts our hypothesis). So hypothesis accepted if H0 is rejected, which is when p-value < alpha. I was mistakenly understand that hypothesis accepted when p-value > alpha.

Thanks to Tara, I get oddity of his inference and thus Im aware of my misconceptions and the bugs [he did not know the concept thus not aware of the bug but he managed to use external tools to take the inference based on p-value. because the inference was factually wrong, but based on the p-value, the inference was true, I noticed that I had the wrong p-value understanding]